Removed IndirectFitOutput
